+++ cypress/support/index.d.ts
@@ -3,6 +3,7 @@ declare namespace Cypress {
   import type { ComponentFixtures } from '../fixtures';
 
   interface Chainable {
+    tab(arg0?: { shift: boolean }): unknown;
     editorEvents(lastN?: number): Chainable<Array<any>>;
     editorActions(lastN?: number): Chainable<Array<any>>;
     setValueExternal(value: any): Chainable<void>;
@@ -10,6 +11,7 @@ declare namespace Cypress {
     mockGoogleMapsResponse(mockData: unknown): void;
     setInitialValue(initialValue: any): void;
     setInitialDisabled(value: boolean | undefined): void;
+    setRestrictedMarks(value: string[]): void;
     setFieldValidations(value: Object[]): void;
     setInstanceParams(value: { [key: string]: any }): void;
     getMarkdownInstance(): Chainable<{