+++ packages/rich-text/src/RichTextEditor.tsx
@@ -29,13 +29,14 @@ type ConnectedProps = {
   onChange?: (doc: Contentful.Document) => unknown;
   isToolbarHidden?: boolean;
   actionsDisabled?: boolean;
+  restrictedMarks?: string[];
 };
 
 export const ConnectedRichTextEditor = (props: ConnectedProps) => {
   const id = getContentfulEditorId(props.sdk);
   const plugins = React.useMemo(
-    () => getPlugins(props.sdk, props.onAction ?? noop),
-    [props.sdk, props.onAction]
+    () => getPlugins(props.sdk, props.onAction ?? noop, props.restrictedMarks),
+    [props.sdk, props.onAction, props.restrictedMarks]
   );
 
   const [isFirstRender, setIsFirstRender] = useState(true);
@@ -119,7 +120,7 @@ export const ConnectedRichTextEditor = (props: ConnectedProps) => {
 type Props = ConnectedProps & { isInitiallyDisabled: boolean };
 
 const RichTextEditor = (props: Props) => {
-  const { sdk, isInitiallyDisabled, onAction, ...otherProps } = props;
+  const { sdk, isInitiallyDisabled, onAction, restrictedMarks, ...otherProps } = props;
   const isEmptyValue = useCallback(
     (value) => !value || deepEquals(value, Contentful.EMPTY_DOCUMENT),
     []
@@ -144,6 +145,7 @@ const RichTextEditor = (props: Props) => {
             onAction={onAction}
             isDisabled={disabled}
             onChange={setValue}
+            restrictedMarks={restrictedMarks}
           />
         )}
       </FieldConnector>