+++ packages/rich-text/src/plugins/Marks/Code.tsx
@@ -15,6 +15,11 @@ export const ToolbarCodeButton = createMarkToolbarButton({
   icon: <CodeIcon />,
 });
 
+export const ToolbarDropdownCodeButton = createMarkToolbarButton({
+  title: 'Code',
+  mark: MARKS.CODE,
+});
+
 const styles = {
   code: css({
     fontFamily: 'monospace',