+++ packages/rich-text/src/plugins/Marks/components/MarkToolbarButton.tsx
@@ -1,6 +1,9 @@
 import * as React from 'react';
 
+import { Menu } from '@contentful/f36-components';
+import tokens from '@contentful/f36-tokens';
 import { MARKS } from '@contentful/rich-text-types';
+import { css, cx } from 'emotion';
 
 import { useContentfulEditor } from '../../../ContentfulEditorProvider';
 import { focus } from '../../../helpers/editor';
@@ -11,8 +14,15 @@ import { toggleMarkAndDeactivateConflictingMarks } from '../helpers';
 export interface MarkOptions {
   mark: MARKS;
   title: string;
-  icon: React.ReactElement;
+  icon?: React.ReactElement;
 }
+
+const styles = {
+  isActive: css({
+    backgroundColor: tokens.blue100,
+    color: tokens.blue600,
+  }),
+};
 
 export const createMarkToolbarButton = ({ mark, title, icon }: MarkOptions) => {
   const Mark = ({ isDisabled }: { isDisabled?: boolean }) => {
@@ -29,6 +39,21 @@ export const createMarkToolbarButton = ({ mark, title, icon }: MarkOptions) => {
     }, [editor]);
 
     if (!editor) return null;
+
+    if (!icon) {
+      return (
+        <Menu.Item
+          onClick={handleClick}
+          disabled={isDisabled}
+          className={cx({
+            [styles.isActive]: isMarkActive(editor, mark),
+          })}
+          testId={`${mark}-toolbar-button`}
+        >
+          {title}
+        </Menu.Item>
+      );
+    }
 
     return (
       <ToolbarButton